requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pandas>=2.2.0
numpy>=2.0.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9
matplotlib>=3.8.2
python-multipart>=0.0.6
pydantic>=2.5.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
requests>=2.31.0
paypalrestsdk>=1.13.3
python-jose[cryptography]>=3.3.0
supabase>=2.0.0
postgrest>=0.13.0

